Goodbye USB, hello Pi
Completely changed the direction of the hardware development.
Originally wanted to utilize a USB interface to plug into any PC but
found it was easier (and cheaper) to by a Raspberry Pi for $35 and build
an expansion board to control the brew hardware.  Created a barebone set
of schematics and layout.