Added 1-wire interface.
The idea is to build a system that can expand as needed by daisy chaning
pieces together.  Instead of having a monolithic controller that has all
of the functionality it will ever need prebuilt into it, keep it a
simple controller but provide some kind of network addressing scheme for
it to commuciate with peripherals.  Commercial HVAC systems use 1-wire
so it should work for our purpose too.  This also means peripherals will
get a little more complicated but the tradeoff should be worth it.